move Code and Contract to genesis.go
